Don't enable plugin if number of workers is 1
If the number of workers requested is 1, then that's equivalent to
running tests serially, so there's no reason to run them in a worker
process with all of the overhead and extra log noise that entails.

Therefore, disable the plugin when the number of requested workers is
1.

Similarly, disable the plugin if the number of workers requested is
the number of processor cores, and there's only one core available.