Better handle tests that don't generate results or take too long
Add a gevented-timeout configuration setting to override the 60-second
timeout for how long we will wait for individual test jobs to
complete.

When the results processor times out, print a list of the tests that
took too long so the user can investigate why.

If the results processor dies or exits because it timed out, stop
sending new tasks to workers, because they're no point in sending
tasks whose results will never be collected.

If the remaining tasks list still has tasks in it when all the worker
processes have exited, then display a warning listing the tasks for
which no results were received, and exit immediately rather than
blocking for 60 seconds (or whatever gevented-timeout is set to).