Add support for timing tests and running them from slowest to fastest
Time how long worker tasks take to execute and debug-log the execution
times.

Add --gevented-timing-file which, if specified, causes the plugin to
store the task execution times for the last test run, and to sort the
tasks from slowest to fastest before executing them the next
time. This increases parallelization and therefore makes test runs
faster.

Incidental change: Collapse TasksQueueManager and TestsQueueManager
into a single class; there was no justification whatsoever that I
could see for keeping the separate TasksQueueManager class, which was
used only by TestsQueueManager, and this artifical separation was
was cluttering the code and making it more complicated.