Save times of tasks over invocations that don't execute them
When a particular test run doesn't execute a particular task, save the
last calculated duration of that task in the gevented-timing-file.

I debated whether to do this, because it means that if a test is
deleted, stale data will persist for it in the state file until the
state file is deleted. But that state takes very little space, and
deleting tests happens much less frequently than running nosetests
over a subset of the tests in a directory, so this seems like a net
gain overall.